Disable include_compiled_code to workaround sqlglot bug
When this setting is true (the default), datahub parses `compiled_code`
from the manifest, and attempts to format it with sqlglot.

However, sqlglot uses recursion in its formatting code, which sometimes breaks
if you have long column expressions.

This is the case for some of our snapshots in CaDeT - they concatenate a
whole lot of columns together as part of the query.

Recursion has been a problem in other parts of sqlglot before,
see https://github.com/tobymao/sqlglot/issues/2961

This setting can be switched back on if the bug is fixed upstream, and
datahub has been upgraded to a fixed version.

It looks like we still need to keep `include_column_lineage` disabled,
as lineage related functionality also depends on sqlglot formatting.

https://github.com/ministryofjustice/find-moj-data/issues/477