fix: default Justice Data domains to "General"
This domain already exists as a catchall for anything not assigned to
another domain, and we think is will be better to assign entities to
General than for an entity to have no domain at all. This way everything
we display in Find MoJ data will have a subject area shown.