fix: downgrade to 0.14.1 of the datahub cli (#322)
fix: downgrade to 0.14.1

I thought 0.14.1.2 was working, but I was testing the wrong job. It seems to
conflict with our create_cadet_databases.yaml ingestion.

All versions past 0.14.1.1 throw an error when pushing
ContainerProperties metadata:

[2024-10-31 09:29:24,753] ERROR    {datahub.ingestion.run.pipeline:77} -  failed to write record with workunit urn:li:container:47cc53f1073124f8906f48248324a4f6-containerProperties with ('Unable to emit metadata to DataHub GMS, likely because the server version is too old relative to the client: Failed to validate record with class com.linkedin.container.ContainerProperties: ERROR :: /env :: unrecognized field found but not allowed\n', {'exceptionClass': 'com.linkedin.restli.server.RestLiServiceException', 'message': 'Failed to validate record with class com.linkedin.container.ContainerProperties: ERROR :: /env :: unrecognized field found but not allowed\n', 'status': 422, 'urn': 'urn:li:container:47cc53f1073124f8906f48248324a4f6', 'workunit_id': 'urn:li:container:47cc53f1073124f8906f48248324a4f6-containerProperties'}) and info {'exceptionClass': 'com.linkedin.restli.server.RestLiServiceException', 'message': 'Failed to validate record with class com.linkedin.container.ContainerProperties: ERROR :: /env :: unrecognized field found but not allowed\n', 'status': 422, 'urn': 'urn:li:container:47cc53f1073124f8906f48248324a4f6', 'workunit_id': 'urn:li:container:47cc53f1073124f8906f48248324a4f6-containerProperties'}

This may be a bug in the python library, as we generate
ContainerProperties indirectly via the `mcp_builder.gen_containers`
helper.

I will investigate further, but for now downgrade to 0.14.1 to unblock ingestion.

This means we cannot use the meta_mapping feature to assign data
custodians, as this has a bug in 0.14.1 (but was fixed in later versions).
I've commented this out temporarily until the issue is resolved.