feat: set subject areas as tags for CaDeT databases
For now, we will dual write a Datahub domain and a datahub tag for the
subject area.

This commit just does CaDeT database, but it will be extended to the rest of the metadata.

Once the tags are fully populated, we'll stop using the domains.

The ultimate goal here is to enable datasets to be tagged to multiple
subject areas, which will give data custodians more flexibility in labelling
their assets, and allow users to discover datasets via multiple paths
through the Find MoJ data browse UI.