perf: reduce search bulk processor limit
In dev, we use the smallest instance size for OpenSearch, which results
in a lot of errors during search indexing, and causes the search index
to become out of date with the database, leading to inconsistencies.

Example error:

OpenSearch exception [type=circuit_breaking_exception, reason=[parent] Data too large, data for [indices:data/write/bulk[s]] would be [1052493344/1003.7mb], which is larger than the limit of [1020054732/972.7mb]...

It looks like we can either increase the heap size available to the
OpenSearch JVM, or we can configure Datahub to not write so many
documents at once.  This commit is an attempt to do the latter.