Remove domain code (#498)
Remove old domain code

Remove everything that adds domain aspect metadata, and renames variables that refer to DataHub domains.

The remaining references to "domain" refer to CaDeT domains - i.e. the domains used in the directory structure of CaDeT.

In addition to removing the domain association in DataHub, this also removes some redundant tags. Before, we were creating tags that exactly matched the CaDeT domains (e.g. "courts", "prison") in addition to the top level domains we actually use ("Courts and tribunals", "Prisons and probation"). We are planning on adding more fine grained subject tags, but instead of having special logic for that here, we will add the tags at source in the CaDeT repo.